+++ Engine/Source/Runtime/Core/Public/Containers/ChunkedArray.h
@@ -1,4 +1,4 @@
-// Copyright 1998-2015 Epic Games, Inc. All Rights Reserved.
+// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
 
 #pragma once
 
@@ -110,7 +110,7 @@ public:
 		return NumElements; 
 	}
 
-	uint32 GetAllocatedSize( void ) const
+	SIZE_T GetAllocatedSize( void ) const
 	{
 		return Chunks.GetAllocatedSize();
 	}
